Improve id generation for subscriptions
Using an incrementing int was causing issues in threaded scenarios which
would result in one subscription getting overwritten. I'm switching to
using a random uuid string and allowing you to also specify an id if
desired. This is a somewhat breaking change since the return type is now
a string.

I also removed the default matcher of exact and made it required to
prevent unexpected behavior.